refactor(linter): Reorganize rules into individual files with new trait system
This refactoring reorganizes lint rules to better support the new Salsa-based architecture:

**Rule Organization:**
- Moved legacy rule implementations from `rules/` to `rules_old/`
- Created new `rules/` directory with one file per rule:
  - `redundant_fields.rs`
  - `operation_name_suffix.rs`
  - `deprecated_field.rs`
  - `require_id_field.rs`
  - `unique_names.rs`
  - `unused_fields.rs`
  - `unused_fragments.rs`

**New Infrastructure:**
- Added `traits.rs` with trait hierarchy:
  - `LintRule` (base trait)
  - `StandaloneDocumentLintRule` (no schema needed)
  - `DocumentSchemaLintRule` (schema access)
  - `ProjectLintRule` (project-wide analysis)
- Added `registry.rs` for centralized rule registration
- Added `diagnostics.rs` with byte-offset based diagnostic types

**Bridge Pattern:**
- New rule implementations delegate to legacy implementations
- Allows gradual migration to HIR queries
- All 95 tests continue to pass

**Next Steps:**
- Migrate individual rules to use HIR queries directly
- Remove legacy implementations once migration is complete
