+++ crates/graphql-linter/src/lib.rs
@@ -1,10 +1,14 @@
 mod config;
 mod context;
 mod linter;
-mod rules;
+
+// Old rule implementations (legacy, to be gradually replaced)
+pub(crate) mod rules_old;
 
 // New Salsa-based architecture
 mod diagnostics;
+mod registry;
+mod rules;
 mod traits;
 
 pub use config::{LintConfig, LintRuleConfig, LintSeverity};
@@ -15,6 +19,7 @@ pub use linter::Linter;
 
 // New architecture exports
 pub use diagnostics::{LintDiagnostic, OffsetRange};
+pub use registry::{document_schema_rules, project_rules, standalone_document_rules};
 pub use traits::{
     DocumentSchemaLintRule, LintRule, ProjectLintRule, StandaloneDocumentLintRule,
     StandaloneSchemaLintRule,
