+++ crates/graphql-linter/src/rules_old/deprecated.rs
