+++ crates/graphql-linter/src/rules_old/field_names_should_be_camel_case.rs
@@ -1,5 +1,5 @@
 use crate::context::StandaloneSchemaContext;
-use crate::rules::StandaloneSchemaRule;
+use crate::rules_old::StandaloneSchemaRule;
 use apollo_compiler::schema::ExtendedType;
 use graphql_project::{Diagnostic, Position, Range, Severity};
